// agent/src/models/stream.ts
export enum ControlType {
    CONTROL_MSG_TYPE_INJECT_KEYCODE,
    CONTROL_MSG_TYPE_INJECT_TEXT,
    CONTROL_MSG_TYPE_INJECT_MOUSE_EVENT,
    CONTROL_MSG_TYPE_INJECT_SCROLL_EVENT,
    CONTROL_MSG_TYPE_BACK_OR_SCREEN_ON,
    CONTROL_MSG_TYPE_EXPAND_NOTIFICATION_PANEL,
    CONTROL_MSG_TYPE_COLLAPSE_NOTIFICATION_PANEL,
    CONTROL_MSG_TYPE_GET_CLIPBOARD,
    CONTROL_MSG_TYPE_SET_CLIPBOARD,
    CONTROL_MSG_TYPE_SET_SCREEN_POWER_MODE,
}
// ...
export interface ControlMsg {
    deviceId: string,
    type: ControlType,
    text: string,
    metaState: number // KeyEvent.META_*
    action: number; // KeyEvent.ACTION_* or MotionEvent.ACTION_* or POWER_MODE_*
    keycode: number; // KeyEvent.KEYCODE_*
    buttons: number; // MotionEvent.BUTTON_*
    pointerId: number;
    pressure: number;
    position: Position;
    hScroll: number;
    vScroll: number;
}

interface Position {
    point: Point,
    screenSize: Size
}

interface Size {
    width: number,
    height: number
}

interface Point {
    x: number,
    y: number,
}